feat(meta): mutable checkpoint frequency (#8010)
As title. Use `LocalNotification` to asyncly notify other components on the meta node of the latest params.

Approved-By: BugenZhao
