+++ src/common/src/system_param/mod.rs
@@ -12,6 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+pub mod reader;
+
 use std::collections::HashSet;
 use std::fmt::Debug;
 use std::ops::RangeBounds;
