+++ src/rpc_client/src/lib.rs
@@ -54,7 +54,7 @@ mod stream_client;
 pub use compute_client::{ComputeClient, ComputeClientPool, ComputeClientPoolRef};
 pub use connector_client::ConnectorClient;
 pub use hummock_meta_client::{CompactTaskItem, HummockMetaClient};
-pub use meta_client::{MetaClient, SystemParamsReader};
+pub use meta_client::MetaClient;
 pub use stream_client::{StreamClient, StreamClientPool, StreamClientPoolRef};
 
 #[async_trait]
